change auth wrapper behaviour

+++ next/src/app/api/auth/logout/route.js
@@ -4,7 +4,9 @@ import Session from '@/lib/server/mongoose/models/Session';
 
 export const POST = async () => {
   const auth = await useServerAuth()
-  await Session.deleteOne({ token: auth.token });
+  if (auth) {
+    await Session.deleteOne({ token: auth.token });
+  }
 
   const res = NextResponse.json({ success: true });
   res.cookies.set('token', '', { path: '/', maxAge: 0 });
